+++ extension/app/scripts.babel/pwa-check.js
@@ -40,7 +40,7 @@ const audits = [
   require('../../../audits/manifest/background-color'),
   require('../../../audits/manifest/theme-color'),
   require('../../../audits/manifest/icons'),
-  require('../../../audits/manifest/icons-192'),
+  require('../../../audits/manifest/icons-min-192'),
   require('../../../audits/manifest/icons-min-144'),
   require('../../../audits/manifest/name'),
   require('../../../audits/manifest/short-name'),
